Extract Gelf decoding logic from codec for reuse
